@@ -8,16 +8,17 @@
     "@fortawesome/fontawesome-svg-core": "^6.4.0",
     "@fortawesome/free-solid-svg-icons": "^6.4.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@mui/icons-material": "^5.11.16",
+    "@mui/icons-material": "^5.14.0",
     "@mui/material": "^5.11.16",
     "@mui/styles": "^5.11.16",
+    "@mui/x-date-pickers": "^6.10.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.22",
-    "@types/react": "^17.0.31",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
     "alert": "^5.1.4",
     "autoprefixer": "^10.4.14",
     "axios": "^1.4.0",
@@ -25,11 +26,13 @@
     "firebase": "^9.22.0",
     "font-awesome": "^4.7.0",
     "postcss": "^8.4.23",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-alert": "^7.0.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.45.2",
+    "react-hook-form-mui": "^6.4.2",
     "react-multi-carousel": "^2.8.4",
-    "react-router-dom": "^6.10.0",
+    "react-router-dom": "^6.14.1",
     "react-scripts": "^5.0.1",
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
@@ -68,8 +71,8 @@
   "devDependencies": {
     "@types/react-alert": "^7.0.2",
     "@types/react-slick": "^0.23.10",
-    "prettier": "^3.0.0",
     "@types/uuid": "^9.0.1",
-    "eslint-plugin-jest": "^27.2.2"
+    "eslint-plugin-jest": "^27.2.2",
+    "prettier": "^3.0.0"
   }
 }
